fix: give buttons a visible focus ring and wire form errors for screen readers
The Button primitive had no visible focus indicator at all. Its focus-visible
ring resolved to a transparent, zero-width shadow, so the app's primary action —
and every other button, since the primitive is shared — failed WCAG 2.4.7. Axe
does not detect this. An explicit outline is used because it cannot be defeated
by the shadow stack; outline-solid is required alongside outline-2 because the
base classes set outline-none, and the width utility alone leaves the style at
none. Measured before and after: outline-style went from `none` to `solid 2px`.

The validation error was announced to nobody: the input carried no aria-invalid
and no aria-describedby, and the error had no id and no role. The count changed
in place after a submit with no live region, so a screen-reader user got no
confirmation the increment worked. Both are now wired, and the pending-writes
notice is a polite live region.

Adds noValidate to the form. The browser's own constraint check ran before
React Hook Form and blocked submission with a bubble in the browser's language
rather than the app's — a decimal was silently dropped that way, with no error,
no request and no feedback. min/max/step remain as spinner affordances. The Zod
schema gains .int(), since the column is an integer and nothing but the browser
had been enforcing that.

Verified in a browser: 1.5 now surfaces the app's own error, aria-invalid and
aria-describedby resolve, and the count is a polite live region.

+++ src/features/counter/schema.ts
@@ -4,7 +4,10 @@ import * as z from 'zod';
  * What the user actually types. The form validates against this.
  */
 export const counterIncrementInputSchema = z.object({
-  increment: z.number().min(1).max(3),
+  // .int() matters: the column is an integer, and without it a fractional
+  // value was stopped only by the browser's implicit step=1 — which the form
+  // now bypasses with noValidate.
+  increment: z.number().int().min(1).max(3),
 });
 
 /**
